Readme.md
Hello, My name is Saksham Mittal. I have made a grocery app under my Android Virtual intership program using Kotlin Language.

As thinking about grocery, it is very difficult to take account of all the vegetables which are present or not. This App will help i keeping list of all the vegetables which are need to be purchase.